Data.Uncertain.Correlated

Description

Provides the Corr monad, which allows one to describe complex relationships between random variables and evaluate their propagated uncertainties respecting their inter-correlations.

Fixes a basic "failure" of the Uncert type, which can't describe correlated samples.

For example, consider the difference between:

sum $ replicate 10 (12.5 +/- 0.8)   -- 125 +/- 3
10 * (12.5 +/- 0.8)                 -- 125 +/- 8

The first one represents the addition of ten independent samples, whose errors will in general cancel eachother out. The second one represents sampling once and multiplying it by ten, which will amplify any error by a full factor of 10.

Documentation

data Corr s a b

evalCorr :: Fractional a => (forall s. Corr s a b) -> b

Evaluates the value described by a Corr monad, taking into account inter-correlations between samples.

Takes a universally qualified Corr, which should not affect usage. See the examples in the documentation for Corr. The univeral qualification is mostly a type system trick to ensure that you aren't allowed to ever use evalCorr to evaluate a CVar.

data CVar :: * -> * -> *

sampleUncert :: Uncert a -> Corr s a (CVar s a)

Generate a sample in Corr from an Uncert value.

sampleExact :: a -> Corr s a (CVar s a)

Generate an exact sample in Corr with zero uncertainty.

Not super useful, since you can do something equivalent with constC or the numeric instances:

sampleExact x  ≡ return (constC x)
sampleExact 10 ≡ return 10

But is provided for completeness alongside sampleUncert.

resolveUncert :: CVar s a -> Corr s a (Uncert a)

Resolve an Uncert from a CVar using its potential multiple samples and sample sources, taking into account inter-correlations between CVars and samples.

Note that if you use sampleUncert on the result, the new sample will be treated as something completely independent. Usually this should only be used as the "exit point" of a Corr description.

liftCF :: (Functor f, Fractional a) => (forall t. f (AD t (Sparse a)) -> AD t (Sparse a)) -> f (CVar s a) -> CVar s a

constC :: a -> CVar s a

liftC :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a

liftC2 :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a -> CVar s a

liftC3 :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a -> CVar s a -> CVar s a

liftC4 :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a -> CVar s a -> CVar s a -> CVar s a

liftC5 :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a -> CVar s a -> CVar s a -> CVar s a -> CVar s a