Data.Uncertain.Correlated.Interactive

Documentation

data Corr s a b

The Corr monad allows us to keep track of correlated and non-independent samples:

evalCorr $ do
  x <- sampleUncert $ 12.5 +/- 0.8
  y1 <- resolveUncert $ sum (replicate 10 x)
  y2 <- resolveUncert $ 10 * x
  return (y1, y2)
  -- result: (125 +- 8, 125 +- 8)

evalCorr $ do
  xs <- replicateM 10 (sampleUncert (12.5 +/- 0.8))
  resolveUncert $ sum xs
  -- result: 125 +/- 3

The first example samples once and describes operations on the single sample; the second example samples 10 times with replicateM and sums all of the results.

Things are more interesting when you sample multiple variables:

evalCorr $ do
  x <- sampleUncert $ 12.5 +/- 0.8
  y <- sampleUncert $ 15.9 +/- 0.5
  z <- sampleUncert $ 1.52 +/- 0.07
  resolveUncert $ (x+z)*logBase z (y**x)
  -- result: 1200 +/- 200

The first parameter is a dummy phantom parameter used to prevent CVars from leaking out of the computation (see evalCorr). The second parameter is the numeric type of all samples within the description (for example, if you ever sample an 'Uncert Double', the second parameter wil be Double). The third parameter is the result type of the computation -- the value the Corr is describing.

runCorrIO :: Corr RealWorld Double a -> IO a

data CVar s a

Represents a single sample (or a value calculated from samples) within the Corr monad. These can be created with sampleUncert, sampleExact, and constC, or made by combinining others with its numeric typeclass instances (like Num) or its functions lifting arbitrary numeric functions (like liftC2). These keep track of inter-correlations between sources, and if you add together two CVars that are correlated, their results will reflect this.

Can be "resolved" into the uncertain value they represent using resolveUncert.

Note that these are parameterized by a dummy phantom parameter s so that they can't be "evaluated" out of the Corr they live in with evalCorr.

Note that a CVar s a can only ever meaningfully "exist" in a Corr s a, meaning that the all samples within that Corr are of the same type.

sampleUncert :: Uncert Double -> IO CVarIO

sampleExact :: Double -> IO CVarIO

resolveUncert :: CVarIO -> IO (Uncert Double)

constC :: a -> CVar s a

Creates a CVar representing a completely independent sample from all other CVars containing the exact value given.

liftC :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a

liftC2 :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a -> CVar s a

liftC3 :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a -> CVar s a -> CVar s a

liftC4 :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a -> CVar s a -> CVar s a -> CVar s a

liftC5 :: Fractional a => (forall t. AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a) -> AD t (Sparse a)) -> CVar s a -> CVar s a -> CVar s a -> CVar s a -> CVar s a -> CVar s a

liftCF :: (Functor f, Fractional a) => (forall t. f (AD t (Sparse a)) -> AD t (Sparse a)) -> f (CVar s a) -> CVar s a

Lifts a multivariate numeric function on a container (given as an f a -> a) to work on a container of CVars. Correctly propagates the uncertainty according to the second-order (multivariate) taylor expansion of the function, and properly takes into account and keeps track of all inter-correlations between the CVar samples. Note that if the higher-degree taylor series terms are large with respect to the means and variances, this approximation may be inaccurate.

Should take any function sufficiently polymorphic over numeric types, so you can use things like *, sqrt, atan2, etc.

evalCorr $ liftCF ([x,y,z] -> x*y+z) [12.2 +- 0.5, 56 +- 2, 0.12 +/- 0.08]
680 +/- 40