Data.Uncertain.Internal

Documentation

data Uncertain a

A data type representing a number with an associated experimental/measurement uncertainty.

So that you can treat it like a normal number, it is an instance of Num, Fractional, Floating, and Real, RealFrac, and RealFloat.

You can apply any function or combination/composition of functions from any of the above typeclasses.

All typeclasses are fully defined, so go crazy! You should technically be able to do anything you can possibly do with a generic instance of Num, Floating, etc.

Some instance notes:

The Show instance displays the "normalized" value (see uNormalize).

The Eq and Ord instances simply compare the central/most likely values.

For RealFrac and RealFloat instances, most functions are defined to be from the central value/most likely value. However, ceiling and floor are taken to include the "entire range":

> ceiling (7.8 +/- 0.4)
9             -- as if calling `ceiling 8.2`
> floor (14.1 +/- 0.3)
12            -- as if calling `floor 13.8`

Constructors

!a :+- !a

uNormalize :: (RealFrac a, Floating a) => Uncertain a -> Uncertain a

Normalize an Uncertain.

This means that the value can't have any more precision than implied by the uncertainty, and that the uncertainty can't have more than one digit of precision/significance.

> uComponents $ uNormalize (15.235812734 +/- 0.0348513)
(15.23, 3.0e-2)

Note that the Show instance of Uncertain automatically normalizes before showing.

This is because, effectively, varying any of the digits that are "normalized away" should yield an indistinguishable Uncertain/experimental result.

Has undefined behavior for "certain" values, with zero uncertainty.

certain :: Num a => a -> Uncertain a

Wraps a Num in an Uncertain, associating it with an uncertainty of 0. Represents a number you know exactly, with no uncertainty.

This is also the default fromInteger method from the Num instance.

> certain 5 * certin 9
certain 45.0
> (18.2 +/- 0.3) ** certain (1/2)
4.271 +/- 4.e-3

Usually, this can be ignored and only the actual literal needs be entered.

> 5 * 9 :: Uncertain Double
certain 45.0
> (18.2 +/- 0.3) ** (1/2)
4.271 +/- 4.e-3

Use when you want an actual function.

> map certain [1,2,3]
[certain 1.0, certain 2.0, certain 3.0]

uVal :: Uncertain a -> a

Get the (central/most likely) value out of an Uncertain.

> uVal (15.2 +/- 0.4)
15.2

uVar :: Uncertain a -> a

Get the uncertainty/root variance out of an Uncertain.

> uVar (15.2 +/- 0.4)
0.4

uComponents :: Uncertain a -> (a, a)

Returns the two components (the central value and the root variance) of an Uncertain.

uComponents = uVal &&& uVar