Data.Uncertain

Documentation

data Uncert a

Represents an independent experimental value centered around a mean value with "inherent" and independent uncertainty.

Mostly useful due to its instances of numeric typeclasses like Num, Fractional, etc., which allows you to add and multiply and apply arbitrary numerical functions to them and have the uncertainty propagate appropriately.

Can be created with exact to represent an "exact" measurement with no uncertainty, +/- and :+/- to specify a standard deviation as a range, withPrecision to specify through decimal precision, and withVar to specify with a variance.

Can be deconstructed with :+/-, the pattern synonym/pseudo-constructor which matches on the mean and a standard deviation. You can also access properties with uMean, uStd, uVar, uMeanStd, uMeanVar, uRange, etc.

It's important to remember that each "occurrence" represents a unique independent sample, so:

> let x = 15 +/- 2 in x + x
30 +/- 3

> let x = 15 +/- 2 in x*2
30 +/- 4

x + x does not represent adding the same sample to itself twice, it represents independently sampling two values within the range 15 +/- 2 and adding them together. In general, errors and deviations will cancel each-other out, leading to a smaller uncertainty.

However, x*2 represents taking one sample and multiplying it by two. This yields a greater uncertainty, because errors and deviations are amplified.

pattern (:+/-) :: () => Floating a => a -> a -> Uncert a infixl 6

uMean :: Uncert a -> a

uVar :: Uncert a -> a

uStd :: Floating a => Uncert a -> a

uMeanVar :: Uncert a -> (a, a)

uMeanStd :: Floating a => Uncert a -> (a, a)

uRange :: Floating a => Uncert a -> (a, a)

(+/-) :: Num a => a -> a -> Uncert a

exact :: Num a => a -> Uncert a

withPrecisionAtBase :: (Floating a, RealFrac a) => Int -> a -> Int -> Uncert a

withPrecision :: (Floating a, RealFrac a) => a -> Int -> Uncert a

withVar :: Num a => a -> a -> Uncert a

uNormalizeAtBase :: (Floating a, RealFrac a) => Int -> Uncert a -> Uncert a

uNormalize :: (Floating a, RealFrac a) => Uncert a -> Uncert a

liftUF :: (Traversable f, Fractional a) => (forall s. f (AD s (Sparse a)) -> AD s (Sparse a)) -> f (Uncert a) -> Uncert a

liftU :: Fractional a => (forall s. AD s (Tower a) -> AD s (Tower a)) -> Uncert a -> Uncert a

liftU' :: Fractional a => (forall s. AD s (Sparse a) -> AD s (Sparse a)) -> Uncert a -> Uncert a

liftU2 :: Fractional a => (forall s. AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a)) -> Uncert a -> Uncert a -> Uncert a

liftU3 :: Fractional a => (forall s. AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a)) -> Uncert a -> Uncert a -> Uncert a -> Uncert a

liftU4 :: Fractional a => (forall s. AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a)) -> Uncert a -> Uncert a -> Uncert a -> Uncert a -> Uncert a

liftU5 :: Fractional a => (forall s. AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a) -> AD s (Sparse a)) -> Uncert a -> Uncert a -> Uncert a -> Uncert a -> Uncert a -> Uncert a